improve initial render()
